pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "portfolioiq"
version = "0.1.0"
description = "Multi-agent financial advisor with bull/bear debates, live news analysis, and ML predictions"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "yfinance>=0.2.0",
    "pandas>=2.0.0",
    "pandas-ta>=0.3.14b",
]

[project.optional-dependencies]
ml = ["scikit-learn>=1.4.0", "xgboost>=2.0.0", "joblib>=1.3.0"]
google = ["langchain-google-genai>=2.0.0"]
anthropic = ["langchain-anthropic>=0.3.0"]
openai = ["langchain-openai>=0.3.0"]
search = ["tavily-python>=0.5.0"]
all = ["portfolioiq[ml,google,anthropic,openai,search]"]

[project.scripts]
portfolioiq = "portfolioiq.__main__:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["portfolioiq*"]
